move memmap_dense_vector_storage
