+++ lib/segment/src/segment_constructor/segment_constructor_base.rs
@@ -33,7 +33,7 @@ use crate::types::{
     VectorStorageType,
 };
 use crate::vector_storage::appendable_mmap_dense_vector_storage::open_appendable_memmap_vector_storage;
-use crate::vector_storage::memmap_dense_vector_storage::open_memmap_vector_storage;
+use crate::vector_storage::dense::memmap_dense_vector_storage::open_memmap_vector_storage;
 use crate::vector_storage::quantized::quantized_vectors::QuantizedVectors;
 use crate::vector_storage::simple_dense_vector_storage::open_simple_vector_storage;
 use crate::vector_storage::simple_sparse_vector_storage::open_simple_sparse_vector_storage;
