+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -1,0 +1,1 @@
+pub mod memmap_dense_vector_storage;
