+++ lib/segment/src/vector_storage/mod.rs
@@ -5,7 +5,6 @@ mod chunked_mmap_vectors;
 mod chunked_utils;
 pub mod chunked_vectors;
 mod dynamic_mmap_flags;
-pub mod memmap_dense_vector_storage;
 mod mmap_dense_vectors;
 pub mod quantized;
 pub mod raw_scorer;
@@ -20,6 +19,7 @@ mod async_io;
 mod async_io_mock;
 mod bitvec;
 pub mod common;
+pub mod dense;
 pub mod query;
 mod query_scorer;
 pub mod simple_multi_dense_vector_storage;
